# src/main/resources/spec/superheroes-openapi-rest.yml
openapi: 3.0.3
info:
  title: Superheroes API
  description: API for CRUD operations on superheroes
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /v1/superheroes:
    get:
      tags:
        - superhero
      summary: Retrieve all superheroes
      description: Returns a list of all superheroes. Allows to search for superheroes containing in their name the value provided in query parameter. For example, if param name is man, the response will be the superheroes Spiderman, Superman, Manolito el fuerte, etc.
      operationId: getSuperheroes
      parameters:
        - name: name
          in: query
          description: Name of the superhero to search for
          example: man
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of superheroes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Superhero'
    post:
      tags:
        - superhero
      summary: Create a superhero
      operationId: createSuperhero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuperheroRequest'
      responses:
        '201':
          description: The created superhero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Superhero'
  /v1/superheroes/{id}:
    get:
      tags:
        - superhero
      summary: Retrieve a superhero by ID
      operationId: getSuperheroById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: The requested superhero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Superhero'
        '404':
          description: Superhero not found
    put:
      tags:
        - superhero
      summary: Update a superhero
      operationId: updateSuperhero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuperheroRequest'
      responses:
        '200':
          description: The updated superhero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Superhero'
        '404':
          description: Superhero not found
    delete:
      tags:
        - superhero
      summary: Delete a superhero
      operationId: deleteSuperhero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: Superhero deleted successfully
        '404':
          description: Superhero not found
components:
  schemas:
    ID:
      type: integer
      format: int64
      example: 1
    Superhero:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          example: "Superman"
    SuperheroRequest:
      type: object
      properties:
        name:
          type: string
          example: "Superman"
      required:
        - name